day 2 + multiply utility

+++ src/main/kotlin/fr/lidonis/adventofcode/y2020/day16/TicketTranslation.kt
@@ -45,7 +45,7 @@ class TicketTranslation(lines: List<String>) {
     }.sum()
 
     fun ticket() = this.fieldsMapping().let { fieldsMapping ->
-        myTicket.mapIndexed { i, value -> fieldsMapping[i] to value }.toMap()
+        myTicket.mapIndexed { i, value -> fieldsMapping[i] to value.toLong() }.toMap()
     }
 
     private fun fieldsMapping(): List<String> {
